feat: move start and end recording to separate functions
